// src/services/types.ts
  export declare type Movie = {
    title: string;
    year: number;
    imdbId: string;
    rank: number;
    actors: string;
    aka: string;
    imdbUrl: string;
    imdbIv: string;
    imgPoster: string;
    photo_width: number;
    photo_height: number;
  };